backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.1.0
click==8.1.2
colorama==0.4.4
fastapi
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
h11==0.13.0
idna==3.3
numpy
pandas
Pillow==9.2.0
pip==21.2.2
protobuf==4.21.4
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
rsa==4.9
selenium==3.141.0
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wheel==0.37.1
wincertstore==0.2
python-jose
passlib
bcrypt
